remember_me, password_reset

+++ db/migrate/20191109195854_create_organisations.rb
@@ -2,7 +2,7 @@ class CreateOrganisations < ActiveRecord::Migration[6.0]
   def change
     create_table :organisations do |t|
       t.string :name
-      t.integer :hourly_rate
+      t.float :hourly_rate
 
       t.timestamps
     end
